feat: Servico update

+++ Servicos/src/main/java/com/SistemaNovaServicos/Servicos/services/ServicoServiceImpl.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 import com.SistemaNovaServicos.Servicos.dto.ServicoSaveDTO;
 import com.SistemaNovaServicos.Servicos.dto.ServicoUpdateDTO;
 import com.SistemaNovaServicos.Servicos.entities.Servico;
+import com.SistemaNovaServicos.Servicos.entities.enums.Estado;
 import com.SistemaNovaServicos.Servicos.repositories.ServicoRepository;
 
 @Service
@@ -39,21 +40,54 @@ public class ServicoServiceImpl implements ServicoService {
     public Servico update(Integer id, ServicoUpdateDTO servicoUpdateDTO) {
         Servico servico = findById(id);
 
+        // o SERVICO so eh atualizado se seu estado for diferente de ENTREGUE ou CANCELADO
+        if (servico.getEstado().equals(Estado.ENTREGUE) || servico.getEstado().equals(Estado.CANCELADO)) {
+            throw new RuntimeException("Impossivel atualizar SERVICO! Motivo: SERVICO" + servico.getEstado());
+        }
+
         servico.setValor(servicoUpdateDTO.valor());
         servico.setDescricao(servicoUpdateDTO.descricao());
 
-        //ADICIONAR CODIGO PARA ATUALIZAR VALOR DO TRABALHO
-        
+        // atualiza o estado se for necessario
+        if (servicoUpdateDTO.estado() != null) {
+            Estado novoEstado = Estado.valueOf(servicoUpdateDTO.estado());
+            servico = updateEstado(servico, novoEstado);
+        }
+
+        // ADICIONAR CODIGO PARA ATUALIZAR VALOR DO TRABALHO
+
         return servicoRepository.save(servico);
     }
 
     @Override
     public Servico delete(Integer id) {
         Servico servico = findById(id);
-        
+
+        if (servico.getEstado().equals(Estado.ENTREGUE)) {
+            throw new RuntimeException("Impossivel excluir: servico com id (" + id + ") ja entregue!");
+        }
+
         servicoRepository.delete(servico);
 
         return servico;
     }
 
+    @Override
+    public Servico deleteAllByTrabalho(Integer id_trabalho) {
+        // ADICIONAR CODIGO PARA DELETAR TODOS COM BASE NO ID DO TRABALHO
+
+        return null;
+    }
+
+    public Servico updateEstado(Servico servico, Estado novoEstado) {
+        if (novoEstado.equals(Estado.ENTREGUE)) {
+            servico.setEstado(servico.getEstado().entregar());
+
+        } else if (novoEstado.equals(Estado.CANCELADO)) {
+            servico.setEstado(servico.getEstado().cancelar());
+
+        }
+
+        return servico;
+    }
 }
